Revert "Move test case to new location"
This reverts commit e2cbdb2b672e1b7a74a94e60e3218c7b79743e88.
